# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-level seat booking system with real-time updates",
        "title": "Seat Booking API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/hold": {
            "post": {
                "description": "Hold a seat for the user's session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hold"
                ],
                "summary": "Hold a seat",
                "parameters": [
                    {
                        "description": "Hold seat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_hold.HoldSeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_hold.HoldStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Release a seat from the user's session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hold"
                ],
                "summary": "Release a held seat",
                "parameters": [
                    {
                        "description": "Release seat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_hold.HoldSeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_hold.HoldStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/hold/extend": {
            "post": {
                "description": "Extend the hold session TTL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hold"
                ],
                "summary": "Extend session",
                "parameters": [
                    {
                        "description": "Extend session request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_hold.ExtendSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_hold.HoldStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/hold/seats": {
            "get": {
                "description": "Get the status of all seats for a show",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hold"
                ],
                "summary": "Get show seats status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show ID",
                        "name": "show_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session ID (optional, to identify own holds)",
                        "name": "session_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_hold.SeatStatusResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/hold/status": {
            "get": {
                "description": "Get the current hold session status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hold"
                ],
                "summary": "Get session status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Show ID",
                        "name": "show_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_hold.HoldStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_shenith404_seat-booking_internal_common.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.ErrorCode"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_shenith404_seat-booking_internal_common.ErrorCode": {
            "type": "string",
            "enum": [
                "INTERNAL_ERROR",
                "BAD_REQUEST",
                "NOT_FOUND",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "CONFLICT",
                "VALIDATION_ERROR",
                "HOLD_SESSION_EXPIRED",
                "HOLD_SESSION_MAX_TIME_EXCEEDED",
                "HOLD_TOGGLE_LIMIT_EXCEEDED",
                "HOLD_SEAT_UNAVAILABLE",
                "HOLD_SEAT_NOT_HELD",
                "BOOKING_SEATS_NOT_HELD",
                "BOOKING_SEATS_ALREADY_BOOKED",
                "BOOKING_FAILED",
                "RATE_LIMIT_EXCEEDED"
            ],
            "x-enum-varnames": [
                "ErrCodeInternal",
                "ErrCodeBadRequest",
                "ErrCodeNotFound",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeConflict",
                "ErrCodeValidation",
                "ErrCodeHoldSessionExpired",
                "ErrCodeHoldSessionMaxTime",
                "ErrCodeHoldToggleLimitHit",
                "ErrCodeHoldSeatUnavailable",
                "ErrCodeHoldSeatNotHeld",
                "ErrCodeBookingSeatsNotHeld",
                "ErrCodeBookingSeatsTaken",
                "ErrCodeBookingFailed",
                "ErrCodeRateLimitExceeded"
            ]
        },
        "github_com_shenith404_seat-booking_internal_common.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_shenith404_seat-booking_internal_common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Error"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_shenith404_seat-booking_internal_common.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_hold.ExtendSessionRequest": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "string"
                },
                "show_id": {
                    "type": "string"
                }
            }
        },
        "internal_hold.HoldSeatRequest": {
            "type": "object",
            "properties": {
                "seat_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "show_id": {
                    "type": "string"
                }
            }
        },
        "internal_hold.HoldStatusResponse": {
            "type": "object",
            "properties": {
                "can_extend": {
                    "type": "boolean"
                },
                "held_seats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_time_seconds": {
                    "type": "integer"
                },
                "remaining_time_seconds": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "show_id": {
                    "type": "string"
                },
                "toggle_count": {
                    "type": "integer"
                }
            }
        },
        "internal_hold.SeatStatusResponse": {
            "type": "object",
            "properties": {
                "seat_id": {
                    "type": "string"
                },
                "session_id": {
                    "description": "Only if held by current session",
                    "type": "string"
                },
                "status": {
                    "description": "\"available\", \"held\", \"booked\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Session-ID",
            "in": "header"
        }
    }
}